+++ package.json
@@ -45,6 +45,7 @@
     "@tsconfig/node18": "^18.2.2",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^20.9.1",
+    "@types/uuid": "^9.0.7",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
     "@typescript-eslint/parser": "^6.11.0",
     "@vitejs/plugin-vue": "^4.2.3",
@@ -63,6 +64,7 @@
     "prettier": "^2.8.8",
     "prettier-config-standard": "^7.0.0",
     "sass": "^1.63.6",
+    "type-fest": "^4.8.3",
     "typescript": "^5.2.2",
     "uuid": "^9.0.1",
     "vite": "^4.3.9",
